criando funçoes
