+++ pokedex/src/pages/homePage.js
@@ -1,36 +1,25 @@
 import React, { useContext, useEffect, useState } from "react"
 import styled from 'styled-components'
-import axios from "axios";
 import {HeaderHome} from '../componetes/headerHome'
 import CardPokemon from "../componetes/CardPokemon";
-import {UrlBase} from '../Constants/Url'
-import GlobalContextState from '../Context/GlobalContextState'
+import GlobalStateContext from "../Context/GlobalContextState";
 
 export const HomePage = () => {
-    
-    const [pokemonsApi, setPokemonsApi] = useState([])
-    const {nome} = useContext(GlobalContextState)
 
-    console.log(nome);
-    const listaPokemons = async () => {
-        try {
-            const response = await axios.get(`${UrlBase}`);
-            setPokemonsApi(response.data.results);
-        } catch (erro) {
-            console.log("Erro", erro);
-        }
-    }
+    const {states, setters, requests} = useContext(GlobalStateContext)
 
     useEffect(() => {
-        listaPokemons();
-    }, [])
+        requests.listaPokemons();
+    }, [requests])
 
-    const listaPokemonsNaTela = pokemonsApi && pokemonsApi.length > 0 && pokemonsApi.map((pokemons)=>{
+    const listaPokemonsNaTela = states.pokemonsApi && states.pokemonsApi.length > 0 && states.pokemonsApi.map((pokemons)=>{
         return <CardPokemon 
                     key={pokemons.name}
                     name={pokemons.name}
                 />
     })
+
+    
 
     return (
         <Principal>
