+++ pokedex/src/pages/pokedexPage.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { useContext, useEffect, useState } from "react"
 import {goToDetalhesPage} from "../routes/coordinator"
 import { useHistory } from "react-router";
 import styled from 'styled-components'
